pyproject.toml
[tool.poetry]
name = "med"
version = "0.1.0"
description = ""
authors = ["Maxim Talancev <maxtalwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
setuptools = "^75.1.0"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
ipython = "^8.28.0"
redis = "^5.1.1"
django-filter = "^24.3"
docker = "^7.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
